@@ -3,18 +3,17 @@
   "version": "0.5.0",
   "private": true,
   "dependencies": {
-    "@blueprintjs/core": "^3.8.0",
-    "@types/react": "^16.7.5",
-    "@types/react-dom": "^16.0.9",
-    "async-es": "^2.6.1",
-    "netlify-cli": "^2.35.0",
+    "@blueprintjs/core": "^5.12.0",
+    "@blueprintjs/icons": "^5.13.0",
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0",
     "normalize.css": "^8.0.1",
-    "react": "^16.6.0",
-    "react-beautiful-dnd": "^10.0.2",
-    "react-device-detect": "^1.6.1",
-    "react-dom": "^16.6.0",
-    "react-scripts": "^2.1.3",
-    "source-map-explorer": "^1.6.0"
+    "react": "^18.3.1",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-device-detect": "^2.2.3",
+    "react-dom": "^18.3.1",
+    "react-scripts": "^5.0.1",
+    "source-map-explorer": "^2.5.3"
   },
   "scripts": {
     "analyze": "source-map-explorer build/static/js/main.*",
@@ -32,5 +31,8 @@
     "not ie <= 11",
     "not op_mini all"
   ],
-  "homepage": "https://www.volumio.org"
+  "homepage": "https://www.volumio.org",
+  "devDependencies": {
+    "jsonwebtoken": "^9.0.3"
+  }
 }
